+++ litellm/proxy/list_api/list_framework.py
@@ -1,4 +1,4 @@
-"""Generic list handling for `/management/v1` collection routes.
+"""Generic list handling for LiteLLM-defined collection routes.
 
 A resource declares a `ListSpec`; `build_query_plan` turns query parameters into a
 `QueryPlan` or an RFC 9457 problem without touching a database, and `handle_list`
@@ -24,7 +24,7 @@ from pydantic import TypeAdapter, ValidationError
 from typing_extensions import assert_never
 
 from litellm.proxy._types import UserAPIKeyAuth
-from litellm.proxy.management_endpoints.management_v1.common import (
+from litellm.proxy.list_api.common import (
     PROBLEM_TYPE_BASE,
     ManagementProblem,
     build_list_links,
