+++ tests/test_litellm/proxy/list_api/test_list_framework.py
@@ -7,13 +7,12 @@ from fastapi import Request
 from pydantic import BaseModel
 
 from litellm.proxy._types import UserAPIKeyAuth
-from litellm.proxy.management_endpoints.management_v1.common import (
-    MANAGEMENT_V1_PREFIX,
+from litellm.proxy.list_api.common import (
     PROBLEM_TYPE_BASE,
     ManagementProblem,
     build_page_links,
 )
-from litellm.proxy.management_endpoints.management_v1.list_framework import (
+from litellm.proxy.list_api.list_framework import (
     AnyOf,
     Compare,
     FilterSpec,
@@ -30,6 +29,7 @@ from litellm.proxy.management_endpoints.management_v1.list_framework import (
     order_by_sql,
     where_sql,
 )
+from litellm.proxy.management_endpoints.management_v1.common import MANAGEMENT_V1_PREFIX
 from litellm.types.proxy.management_endpoints.management_v1 import (
     PageLinks,
     PageMeta,
